requirements.txt
playwright>=1.44.0
python-dotenv>=1.0.1
SQLAlchemy>=2.0.29
PyMySQL>=1.1.1
alembic>=1.13.1
cryptography>=43.0.0
bcrypt>=4.1.3
